ragui/requirements.txt
streamlit
langchain
langchain_community
langchain_core
sentence_transformers
elasticsearch
python-dotenv
langchain_openai